allowed using float values for tiles

+++ core/modules/tile/internal/tilerenderer/systemRegister.go
@@ -23,7 +23,7 @@ import (
 )
 
 type TileData struct {
-	PosX, PosY int32
+	PosX, PosY float32
 	Type       definition.DefinitionID
 }
